fix: multiple default signature error,sign-yourself placeholder location issue in mobile view
